tuber-ecs: Move query related code to a separate module
