+++ crates/tuber-ecs/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod ecs;
+pub mod query;
 
 /// The index of an entity
 pub type EntityIndex = usize;
